refactor: change view place
