+++ apps/web/src/views/FarmPoolDetail/FarmPoolDetail.tsx
