+++ apps/web/src/views/FarmPoolDetail/index.ts
@@ -1,0 +1,1 @@
+export * from './FarmPoolDetail'
